+++ lib/app_sliver_stacked.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class AppSliverStacked extends StatelessWidget {
+  const AppSliverStacked({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}